// src/types/proposals.ts
export type ProposalType = {
  title: string;
  chain: string;
  dateCreated: string;
  type: string;
  bond: string;
  reward: string;
  request: string[];
  description: string;
  string: string;
  oracle: string;
  umip: string;
  identifier: string;
  requester: string;
  requestTxn: string;
};